# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport Api Swagger Documentation Anatation",
        "description": "Sport Api Documentation to build the frontend of this site",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "abduvaliyevasad23@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Sport api server link"
        }
    ],
    "paths": {
...
        "/api/training": {
            "get": {
                "tags": [
                    "Training"
                ],
                "summary": "Get all data from training database",
                "description": "Via this link All training` datas come",
                "operationId": "training",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Training"
                ],
                "summary": "Register a training",
                "operationId": "storeTraining",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Type user id",
                                        "type": "integer"
                                    },
                                    "region_id": {
                                        "description": "Type region id",
                                        "type": "integer"
                                    },
                                    "speciality": {
                                        "description": "Type fillial id",
                                        "type": "string"
                                    },
                                    "direction": {
                                        "description": "Type your direction",
                                        "type": "string"
                                    },
                                    "date_start": {
                                        "description": "Type your start date",
                                        "type": "string"
                                    },
                                    "date_end": {
                                        "description": "Type your end date",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/training/{training}": {
            "get": {
                "tags": [
                    "Training"
                ],
                "summary": "Get one data from Training database",
                "description": "Via this link a Training`s data comes to show",
                "operationId": "showTraining",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "training",
                        "in": "path",
                        "description": "ID for training data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/training/{training}/update": {
            "put": {
                "tags": [
                    "Training"
                ],
                "summary": "Updated",
                "description": "Update this training",
                "operationId": "updateTraining",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "training",
                        "in": "path",
                        "description": "training that to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Type user id",
                                        "type": "integer"
                                    },
                                    "region_id": {
                                        "description": "Type region id",
                                        "type": "integer"
                                    },
                                    "speciality": {
                                        "description": "Type fillial id",
                                        "type": "string"
                                    },
                                    "direction": {
                                        "description": "Type your direction",
                                        "type": "string"
                                    },
                                    "date_start": {
                                        "description": "Type your start date",
                                        "type": "string"
                                    },
                                    "date_end": {
                                        "description": "Type your end date",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid training supplied"
                    },
                    "404": {
                        "description": "training not found"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/training/{training}/delete": {
            "delete": {
                "tags": [
                    "Training"
                ],
                "summary": "Deletes a training",
                "operationId": "destroyTraining",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "training",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid course supplied"
                    },
                    "404": {
                        "description": "course not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}